backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
MetaTrader5==5.0.45
python-dotenv==1.0.0
cryptography==41.0.7
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
